+++ src/commands/index.ts
@@ -1,2 +1,3 @@
 export * from './helloWorld';
 export * from './installUpdate';
+export * from './pickVersion';