+++ src/utils/index.ts
@@ -2,6 +2,7 @@ export * from './packageReader';
 export * from './ncuClient';
 export * from './packageManager';
 export * from './versionUtils';
+export * from './registryClient';
 export * from './notify';
 export * from './logger';
 export * from './auditClient';